Move upload button to send box toolbar middleware
