+++ packages/component/src/SendBoxToolbar/Assets/AttachmentIcon.tsx
