+++ packages/component/src/SendBoxToolbar/BasicSendBoxToolbar.tsx
@@ -1,0 +1,14 @@
+import { hooks, type SendBoxToolbarMiddlewareProps } from 'botframework-webchat-api';
+import React, { memo } from 'react';
+
+import UploadButton from './UploadButton';
+
+const { useStyleOptions } = hooks;
+
+function BasicSendBoxToolbar({ className }: SendBoxToolbarMiddlewareProps) {
+  const [{ hideUploadButton }] = useStyleOptions();
+
+  return !hideUploadButton && <UploadButton className={className} />;
+}
+
+export default memo(BasicSendBoxToolbar);
