+++ packages/component/src/SendBoxToolbar/createMiddleware.tsx
@@ -1,0 +1,7 @@
+import { type SendBoxMiddleware } from 'botframework-webchat-api';
+
+import BasicSendBoxToolbar from './BasicSendBoxToolbar';
+
+const createMiddleware = (): readonly SendBoxMiddleware[] => Object.freeze([() => () => () => BasicSendBoxToolbar]);
+
+export default createMiddleware;
